+++ libs/i18n/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.51.3-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.6...v0.51.3-rc.7) (2024-08-12)
+
+**Note:** Version bump only for package @fundamental-ngx/i18n
+
+
+
+
+
+## [0.51.3-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.5...v0.51.3-rc.6) (2024-08-12)
+
+**Note:** Version bump only for package @fundamental-ngx/i18n
+
+
+
+
+
+## [0.51.3-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.4...v0.51.3-rc.5) (2024-08-07)
+
+**Note:** Version bump only for package @fundamental-ngx/i18n
+
+
+
+
+
+## [0.51.3-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.3...v0.51.3-rc.4) (2024-08-06)
+
+**Note:** Version bump only for package @fundamental-ngx/i18n
+
+
+
+
+
+## [0.51.3-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.2...v0.51.3-rc.3) (2024-08-06)
+
+**Note:** Version bump only for package @fundamental-ngx/i18n
+
+
+
+
+
 ## [0.51.3-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.1...v0.51.3-rc.2) (2024-08-04)
 
 **Note:** Version bump only for package @fundamental-ngx/i18n